Custom Chatbot Using LangChain and gemini/requirements.txt
InstructorEmbedding==1.0.1
langchain==0.3.3
protobuf==5.28.2
pydantic==2.9.2
python-dotenv==1.0.1
streamlit==1.39.0
